Moved collections to pages folder
