+++ client/src/components/Dashboard/index.js
@@ -3,7 +3,7 @@ import {Link} from 'react-router-dom'
 import PropTypes from 'prop-types'
 import { makeStyles } from '@material-ui/core/styles'
 import { AppBar, Tabs, Tab, Typography, Box } from '@material-ui/core'
-import Collection from '../Collection'
+import Collection from '../../pages/Collection'
 import Packs from '../../pages/Packs'
 import ProfileComponent from '../Profile'
 import { AuthContext } from '../../Context/AuthContext'
