+++ client/src/pages/Collection.js
@@ -1,9 +1,9 @@
 import React, { useContext, useEffect, useState } from 'react';
 import { makeStyles } from '@material-ui/styles';
-import CardFlip from '../../components/CardFlip';
+import CardFlip from '../components/CardFlip';
 // import ItemsCarousel from 'react-items-carousel';
-import API from '../../utils/API';
-import { AuthContext } from '../../Context/AuthContext'
+import API from '../utils/API';
+import { AuthContext } from '../Context/AuthContext'
 
 // import Swiper from 'react-id-swiper';
 import 'swiper/css/swiper.css'
